// iuffimobile/src/app/detections/api/get-photos-response.ts
export interface PhotoItemResponse {
	idFoto: number;
	foto?: any;
	fotoByte?: any;
	nomeFile: string;
	dataFoto: number;
	tag?: any;
	note?: any;
	base64: string;
	anno?: any;
	dallaData?: any;
	allaData?: any;
	dallaDataS?: any;
	allaDataS?: any;
	visual?: any;
	trappolaggio?: any;
	campionamento?: any;
	numeroTrasferta?: any;
	tipologia?: any;
	tipoArea?: any;
	ispettoreAssegnato?: any;
	ispettoriSecondari?: any;
	specieVegetale?: any;
	organismoNocivo?: any;
	trappole?: any;
	campioni?: any;
	istatComune?: any;
	specie?: any;
	area?: any;
	trappola?: any;
	campione?: any;
	longitudine: number;
	latitudine: number;
	ispettoreEvento?: any;
	ispettoreAssegnatoM?: any;
	descrComune?: any;
	organismi?: any;
	idVisual: number;
	idCampionamento: number;
	idTrappolaggio: number;
	dataFotoS: string;
	dataFotoF: string;
}