// iuffimobile/src/app/detections/dto/trap.dto.ts
export interface TrapDto {
    idTrappola: number;
    codiceSfr: string;
    dataInstallazione: string;
    dataRimozione: string;
    idSpecieVeg: number;
    specie: string;
    idTipoTrappola: number;
    idOrganismoNocivo: number;
    latitudine: number;
    longitudine: number;
    descrTrappola: string;
}